add firebase to our project &create app router&cubits and some screens

+++ lib/models/note_model.dart
@@ -14,7 +14,7 @@ class NoteModel extends HiveObject {
   final String date;
 
   @HiveField(3)
-  int color;
+  double color;
 
   NoteModel({
     required this.title,
